Move different identifier's types to separate classes
