+++ lib/pubid/iso.rb
@@ -19,7 +19,15 @@ require_relative "iso/transformer"
 require_relative "iso/supplement"
 require_relative "iso/amendment"
 require_relative "iso/corrigendum"
-require_relative "iso/identifier"
+require_relative "iso/identifier/base"
+require_relative "iso/identifier/international_standard"
+require_relative "iso/identifier/technical_report"
+require_relative "iso/identifier/technical_specification"
+require_relative "iso/identifier/supplement"
+require_relative "iso/identifier/amendment"
+require_relative "iso/identifier/corrigendum"
+require_relative "iso/identifier/publicly_available_specification"
+require_relative "iso/identifier/directives"
 require_relative "iso/renderer/base"
 require_relative "iso/renderer/urn"
 require_relative "iso/renderer/french"
@@ -28,3 +36,8 @@ require_relative "iso/renderer/tc"
 require_relative "iso/renderer/urn-tc"
 require_relative "iso/renderer/dir"
 require_relative "iso/renderer/urn-dir"
+require_relative "iso/renderer/technical_report"
+require_relative "iso/renderer/supplement"
+require_relative "iso/renderer/amendment"
+require_relative "iso/renderer/corrigendum"
+require_relative "iso/renderer/publicly_available_specification"
