+++ lib/pubid/iso/identifier/technical_specification.rb
@@ -1,0 +1,24 @@
+module Pubid::Iso
+  module Identifier
+    class TechnicalSpecification < Base
+      def_delegators 'Pubid::Iso::Identifier::TechnicalSpecification', :type
+
+      TYPED_STAGES = {
+        dts: {
+          abbr: "DTS",
+          name: "Draft Technical Specification",
+          harmonized_stages: %w[40.00 40.20 40.60 40.92 40.93 50.00 50.20 50.60 50.92],
+        },
+        fdts: {
+          abbr: "FDTS",
+          name: "Final Draft Technical Specification",
+          harmonized_stages: %w[50.00 50.20 50.60 50.92],
+        },
+      }.freeze
+
+      def self.type
+        :ts
+      end
+    end
+  end
+end
