+++ lib/pubid/iso/renderer/amendment.rb
@@ -1,0 +1,4 @@
+module Pubid::Iso::Renderer
+  class Amendment < Supplement
+  end
+end
