+++ lib/pubid/iso/renderer/corrigendum.rb
@@ -1,0 +1,4 @@
+module Pubid::Iso::Renderer
+  class Corrigendum < Supplement
+  end
+end
