+++ lib/pubid/iso/typed_stage.rb
@@ -3,72 +3,6 @@ module Pubid::Iso
     attr_accessor :type, :stage, :typed_stage
 
     TYPED_STAGES = {
-      dtr: {
-        abbr: "DTR",
-        type: :tr,
-        name: "Draft Technical Report",
-        harmonized_stages: %w[40.00 40.20 40.60 40.92 40.93 50.00 50.20 50.60 50.92],
-      },
-      dis: {
-        abbr: "DIS",
-        type: :is,
-        name: "Draft International Standard",
-        harmonized_stages: %w[40.00 40.20 40.60 40.92 40.93],
-      },
-      dts: {
-        abbr: "DTS",
-        type: :ts,
-        name: "Draft Technical Specification",
-        harmonized_stages: %w[40.00 40.20 40.60 40.92 40.93 50.00 50.20 50.60 50.92],
-      },
-      fdts: {
-        abbr: "FDTS",
-        type: :ts,
-        name: "Final Draft Technical Specification",
-        harmonized_stages: %w[50.00 50.20 50.60 50.92],
-      },
-      fdtr: {
-        abbr: "FDTR",
-        type: :tr,
-        name: "Final Draft Technical Report",
-        harmonized_stages: %w[50.00 50.20 50.60 50.92],
-      },
-      fdis: {
-        abbr: "FDIS",
-        type: :is,
-        name: "Final Draft International Standard",
-        harmonized_stages: %w[50.00 50.20 50.60 50.92],
-      },
-      dpas: {
-        abbr: "DPAS",
-        type: :pas,
-        name: "Publicly Available Specification Draft",
-        harmonized_stages: %w[40.00 40.20 40.60 40.92 40.93 50.00 50.20 50.60 50.92],
-      },
-      damd: {
-        abbr: { short: "DAM", long: "DAmd" },
-        type: :amd,
-        name: "Draft Amendment",
-        harmonized_stages: %w[40.00 40.20 40.60 40.92 40.93 50.00 50.20 50.60 50.92],
-      },
-      dcor: {
-        abbr: { short: "DCOR", long: "DCor" },
-        type: :cor,
-        name: "Draft Corrigendum",
-        harmonized_stages: %w[40.00 40.20 40.60 40.92 40.93 50.00 50.20 50.60 50.92],
-      },
-      fdamd: {
-        abbr: { short: "FDAM", long: "FDAmd" },
-        type: :amd,
-        name: "Final Draft Amendment",
-        harmonized_stages: %w[50.00 50.20 50.60 50.92],
-      },
-      fdcor: {
-        abbr: { short: "FDCOR", long: "FDCor" },
-        type: :cor,
-        name: "Final Draft Corrigendum",
-        harmonized_stages: %w[50.00 50.20 50.60 50.92],
-      },
     }.freeze
 
     # @param type [Symbol,Type] eg. :tr, Type.new(:tr)
